Remove status_monitor module
